+++ MochiKit/Logging.js
@@ -335,15 +335,20 @@ MochiKit.Logging.__new__ = function () {
     Logger.prototype.fatal = partial(Logger.prototype.baseLog, 'FATAL');
     Logger.prototype.warning = partial(Logger.prototype.baseLog, 'WARNING');
 
-    var bind = MochiKit.Base.bind;
+    // indirectly find logger so it can be replaced
+    var self = this;
+    var connectLog = function (name) {
+        return function () {
+            self.logger[name].apply(self.logger, arguments);
+        }
+    };
 
-    var logger = new Logger();
-    this.log = bind(logger.log, logger);
-    this.logError = bind(logger.error, logger);
-    this.logDebug = bind(logger.debug, logger);
-    this.logFatal = bind(logger.fatal, logger);
-    this.logWarning = bind(logger.warning, logger);
-    this.logger = logger;
+    this.log = connectLog('log');
+    this.logError = connectLog('error');
+    this.logDebug = connectLog('debug');
+    this.logFatal = connectLog('fatal');
+    this.logWarning = connectLog('warning');
+    this.logger = new Logger();
 
     this.EXPORT_TAGS = {
         ":common": this.EXPORT,